Now, Crossdata is launched as an application within the docker and not as an systemd
Now, Crossdata will be launched as an application within the docker, not as an INIT daemon. That makes its output to be correctly redirected to cluster managers as well as the docker container be stopped if the server crashes.

If the docker is run with the option -e SERVER_MODE=debug it will launch Crossdata as a service within the docker container, that is, as before the PR was merged.